refactor search-intolerances as a component
